Prepare to support parallel behat runs
Although this functionality does not currently store any data, it may do
in the future.

To prepare for this, we should add a serverID into the URI so that a
unique id can be used per parallel run.

Typically we've done this in the past with a md5/shasum of the
behat_wwwroot.